JSON Powered Polish list display.

+++ PolishWarehouse/Models/BrandModel.cs
@@ -5,7 +5,7 @@ namespace PolishWarehouse.Models
 {
     public class BrandModel
     {
-        public long? ID { get; set; }
+        public int? ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public int? CategoryID { get; set; }
