+++ PolishWarehouse/Models/ColorModel.cs
@@ -5,7 +5,7 @@ namespace PolishWarehouse.Models
 {
     public class ColorModel
     {
-        public long? ID { get; set; }
+        public int? ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public bool IsPrimary { get; set; }
